+++ core/config/envvar/schema.go
@@ -167,20 +167,20 @@ type ConfigSchema struct {
 	EvmGasBumpThreshold   uint64   `env:"ETH_GAS_BUMP_THRESHOLD"`
 	EvmGasBumpWei         *big.Int `env:"ETH_GAS_BUMP_WEI"`
 	EvmGasFeeCapDefault   *big.Int `env:"EVM_GAS_FEE_CAP_DEFAULT"`
-	EvmGasLimitDefault    uint64   `env:"ETH_GAS_LIMIT_DEFAULT"`
+	EvmGasLimitDefault    uint32   `env:"ETH_GAS_LIMIT_DEFAULT"`
 	EvmGasLimitMultiplier float32  `env:"ETH_GAS_LIMIT_MULTIPLIER"`
-	EvmGasLimitTransfer   uint64   `env:"ETH_GAS_LIMIT_TRANSFER"`
+	EvmGasLimitTransfer   uint32   `env:"ETH_GAS_LIMIT_TRANSFER"`
 	EvmGasPriceDefault    *big.Int `env:"ETH_GAS_PRICE_DEFAULT"`
 	EvmGasTipCapDefault   *big.Int `env:"EVM_GAS_TIP_CAP_DEFAULT"`
 	EvmGasTipCapMinimum   *big.Int `env:"EVM_GAS_TIP_CAP_MINIMUM"`
 	EvmMaxGasPriceWei     *big.Int `env:"ETH_MAX_GAS_PRICE_WEI"`
 	EvmMinGasPriceWei     *big.Int `env:"ETH_MIN_GAS_PRICE_WEI"`
 	// Gas limits per job type
-	EvmGasLimitOCRJobType    *uint64 `env:"ETH_GAS_LIMIT_OCR_JOB_TYPE"`
-	EvmGasLimitDRJobType     *uint64 `env:"ETH_GAS_LIMIT_DR_JOB_TYPE"`
-	EvmGasLimitVRFJobType    *uint64 `env:"ETH_GAS_LIMIT_VRF_JOB_TYPE"`
-	EvmGasLimitFMJobType     *uint64 `env:"ETH_GAS_LIMIT_FM_JOB_TYPE"`
-	EvmGasLimitKeeperJobType *uint64 `env:"ETH_GAS_LIMIT_KEEPER_JOB_TYPE"`
+	EvmGasLimitOCRJobType    *uint32 `env:"ETH_GAS_LIMIT_OCR_JOB_TYPE"`
+	EvmGasLimitDRJobType     *uint32 `env:"ETH_GAS_LIMIT_DR_JOB_TYPE"`
+	EvmGasLimitVRFJobType    *uint32 `env:"ETH_GAS_LIMIT_VRF_JOB_TYPE"`
+	EvmGasLimitFMJobType     *uint32 `env:"ETH_GAS_LIMIT_FM_JOB_TYPE"`
+	EvmGasLimitKeeperJobType *uint32 `env:"ETH_GAS_LIMIT_KEEPER_JOB_TYPE"`
 	// Gas Estimation
 	GasEstimatorMode                               string `env:"GAS_ESTIMATOR_MODE"`
 	BlockHistoryEstimatorBatchSize                 uint32 `env:"BLOCK_HISTORY_ESTIMATOR_BATCH_SIZE"`
